refactor(toast, modal): move commom functionality into custom hook
