+++ src/hooks/__internal__/useModalManager/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./useModalManager";
